--- src/components/Desk.tsx
@@ -1,4 +1,0 @@
-const Desk: React.FC = () => {
-  return <div></div>;
-};
-export default Desk;
