+++ src/components/blocks/studio/Preview/index.tsx
@@ -1,0 +1,12 @@
+import PlayButton from "#/components/blocks/studio/Preview/PlayButton";
+import Player from "#/components/blocks/studio/Preview/Player";
+
+const Preview = () => {
+  return (
+    <div>
+      <Player />
+      <PlayButton />
+    </div>
+  );
+};
+export default Preview;
