+++ src/pages/project/[pid]/studio.tsx
@@ -1,40 +1,14 @@
-import { Button } from "@blueprintjs/core";
-import { useRouter } from "next/router";
-
-import CreateNewObject from "#/components/CreateNewObject";
-import PlayButton from "#/components/PlayButton";
-import Player from "#/components/Player";
-import ProjectName from "#/components/ProjectName";
-import Toolbar from "#/components/Toolbar";
+import Header from "#/components/blocks/studio/Header";
+import CreateNewObject from "#/components/blocks/studio/ObjectConfig/CreateNewObject";
+import Preview from "#/components/blocks/studio/Preview";
 
 const Studio = () => {
-  const router = useRouter();
-
   return (
     <div>
       <main>
-        <div
-          style={{
-            display: "flex",
-            flexDirection: "column",
-            gap: "1rem",
-          }}
-        >
-          <div
-            style={{
-              display: "flex",
-              alignItems: "center",
-              gap: "1rem",
-            }}
-          >
-            <Button icon="home" onClick={() => router.push("/")} />
-            <ProjectName />
-          </div>
-          <Toolbar />
-          <Player />
-          <PlayButton />
-          <CreateNewObject />
-        </div>
+        <Header />
+        <Preview />
+        <CreateNewObject />
       </main>
     </div>
   );
